requirements.txt
python-dotenv==1.0.1
Django==5.0.3
psycopg==3.1.18
djangorestframework==3.15.1